Création d'une nouvelle table role dans la base pour gérer l'accès aux pages du site selon le rôle d'un utilisateur. Ainsi un utilisateur admin pourra accéder à toutes les pages et avoir un contrôle sur tout les éléments du site avec une page Administration qui lui est réservé. Un utilisateur User pourra lui seulement consulter les articles, son profil, ajouter des commentaires, signaler des commentaires.Pour avoir cette séparation sur le site, j'ai simplement créer une nouvelle function checkUserRole dans mon UserModel pour donc vérifier le rôle d'un utilisateur. Cette fonction est appelée à chaque fois qu'un utlisateur se connecte (dans la function login). Dans cette function login, je stock aussi certaines informations de l'utilisateur en SESSION qui me seront utiles par la suite, par exemple pour avoir directement le pseudo de l'utilisateur lorsque qu'il ajoutera un nouveau commentaire. Je stock aussi le role pour effectuer une condition dessus dans ma vue principale pour afficher ou non le lien de la page d'Administration. Parmi les autres améliorations apportées, j'ai empêcher l'accès aux pages en faisant des redirections selon connection role de l'utilisateur pour ainsi éviter de permettre de passer dans les routes manuellement dans l'url du site. Prochaine amélioration : Bloquer la page d'administration par un mot de passe.

+++ controller/ArticleController.php
@@ -20,36 +20,42 @@ class ArticleController extends Controller{
     }
 
     public function addArticle($post){
-        if(isset($post['submit'])){
-            $errors = $this->validation->validate($post, 'Article');
-            if(!($errors)){
-                $this->articleModel->addArticle($post);
-                $this->session->set('add_article', 'Le nouvel article a bien été ajouté');
-                header('Location: ../public/index.php');
+        if($this->checkAdmin()){
+            if(isset($post['submit'])){
+                $errors = $this->validation->validate($post, 'Article');
+                if(!($errors)){
+                    $this->articleModel->addArticle($post);
+                    $this->session->set('add_article', 'Le nouvel article a bien été ajouté');
+                    header('Location: ../public/index.php');
+                }
+                return $this->view->render('CreateArticle', [
+                    'post' => $post,
+                    'errors' => $errors
+                ]);
             }
-            return $this->view->render('CreateArticle', [
-                'post' => $post,
-                'errors' => $errors
-            ]);
+            return $this->view->render('CreateArticle');
         }
-        return $this->view->render('CreateArticle');
     }
 
     public function updateArticle($post, $articleId){
-        $article = $this->articleModel->getArticle($articleId);
-        if(isset($post['submit'])){
-            $this->articleModel->updateArticle($post, $articleId);
-            $this->session->set('update_article', 'L\'article a bien été modifié');
-            header('Location: ../public/index.php?route=Administration');
+        if($this->checkAdmin()){
+            $article = $this->articleModel->getArticle($articleId);
+            if(isset($post['submit'])){
+                $this->articleModel->updateArticle($post, $articleId);
+                $this->session->set('update_article', 'L\'article a bien été modifié');
+                header('Location: ../public/index.php?route=Administration');
+            }
+            return $this->view->render('UpdateArticle', [
+                'article'=>$article
+            ]);
         }
-        return $this->view->render('UpdateArticle', [
-            'article'=>$article
-        ]);
     }
 
     public function deleteArticle($articleId){
-        $this->articleModel->deleteArticle($articleId);
-        $this->session->set('delete_article', 'L\'article a bien été supprimé');
-        header('Location: ../public/index.php?route=Administration');
+        if($this->checkAdmin()){
+            $this->articleModel->deleteArticle($articleId);
+            $this->session->set('delete_article', 'L\'article a bien été supprimé');
+            header('Location: ../public/index.php?route=Administration');
+        }
     }
 }
